Part 1 of the move to a new name: SEIZMO This will avoid the confusion with Mike Thorne's SAClab that this originally started from.  The name SEIZMO basically means SEIZmology for Matlab and Octave.
There are number of new additions too as we make the move:
1 - starting to use Matlab's Contents.m system to make the
    full use of the help system
2 - date and time functions for handling just about every
    situation I can think of for now - particularly this
    allows for leap second awareness
3 - slidingavg takes the place of slidingmean but is now
    just a basic function (ie works on numeric arrays) and
    has options to slide across any dimension and also 
    includes a custom average option.
4 - submat* pulled in from mcxc package - the rest will be
    pulled in soon too
5 - merge, rotate, convolve, correlate are not done or not
    even started but they will be the new functions of the
    next release
6 - getapp allows detection of if the user is running Matlab
    or Octave and also what version they are using

I can't think of any more at the moment (there are more though)

There are lots of renames at the moment as I change names of
functions to be a bit longer so they are more understandable.
This brings massive breakage.  This revision will not function
for the most part.  Also some of the names will conflict with
more obscure Matlab toolboxes (curve fitting has a integrate
and a differentiate - another has merge).  For now I prefer
to have these longer names with some conflict rather than have
to come up with unique ones that are hard to remember.  Along
this line I am planning to make have a shortcuts folder that
will allow users to have the short names for functions too
so everything works sorta like SAC.  I'm still working out how
to implement this (it will be a folder of links but it how to
deal with windows vs unix is the main problem along with how
to include the links (an install script likely).