Name change part 2:
Should not be any SAClab/SACLAB left now.

Lots of function renaming too.  Mainly towards longer
and more descriptive names.

All private functions went through an update, so they
should be working now.  One interesting add in the
update was caching for seizmodef - gave an order of
magnitude speed increase for that function.  Pretty
good considering it is the kernel of this toolbox.
Also separated SAC from SEIZMO, which should force
me to clean up code here and there.

As far as main functions go...some are now working.
Not many though.  Also .endian is now .byteorder .
And sending 'dep' to seizmocheck will cause it to
require that .hasdata is TRUE.  This will make data
dependent functions a bit easier to write (still does
not catch npts==0). There are a lot of name changes 
but they should be easy to figure out.
Hardest:
 - subsample == deci
 - oversample == stretch